Temporally keep file test/hotspot/gtest/aarch64/asmtest.out.h same with
branch mastger.

+++ src/jdk.hotspot.agent/share/classes/sun/jvm/hotspot/oops/DataLayout.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011, 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2011, 2021, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -81,9 +81,8 @@ public class DataLayout {
     return data.getJShortAt(offset + at) & 0xffff;
   }
 
-  int cellAt(int index) {
-    // Cells are intptr_t sized but only contain ints as raw values
-    return (int)data.getCIntegerAt(offset + cellOffset(index), MethodData.cellSize, false);
+  long cellAt(int index) {
+    return data.getCIntegerAt(offset + cellOffset(index), MethodData.cellSize, false);
   }
 
   public Address addressAt(int index) {
