+++ src/jdk.incubator.foreign/share/classes/jdk/internal/foreign/HeapMemorySegmentImpl.java
@@ -98,7 +98,7 @@ public abstract class HeapMemorySegmentImpl<H> extends AbstractMemorySegmentImpl
 
         public static MemorySegment fromArray(byte[] arr) {
             Objects.requireNonNull(arr);
-            int byteSize = arr.length * Unsafe.ARRAY_BYTE_INDEX_SCALE;
+            long byteSize = (long)arr.length * Unsafe.ARRAY_BYTE_INDEX_SCALE;
             MemoryScope scope = MemoryScope.createConfined(null, MemoryScope.DUMMY_CLEANUP_ACTION, null);
             return new OfByte(Unsafe.ARRAY_BYTE_BASE_OFFSET, arr, byteSize, defaultAccessModes(byteSize), scope);
         }
@@ -122,7 +122,7 @@ public abstract class HeapMemorySegmentImpl<H> extends AbstractMemorySegmentImpl
 
         public static MemorySegment fromArray(char[] arr) {
             Objects.requireNonNull(arr);
-            int byteSize = arr.length * Unsafe.ARRAY_CHAR_INDEX_SCALE;
+            long byteSize = (long)arr.length * Unsafe.ARRAY_CHAR_INDEX_SCALE;
             MemoryScope scope = MemoryScope.createConfined(null, MemoryScope.DUMMY_CLEANUP_ACTION, null);
             return new OfChar(Unsafe.ARRAY_CHAR_BASE_OFFSET, arr, byteSize, defaultAccessModes(byteSize), scope);
         }
@@ -146,7 +146,7 @@ public abstract class HeapMemorySegmentImpl<H> extends AbstractMemorySegmentImpl
 
         public static MemorySegment fromArray(short[] arr) {
             Objects.requireNonNull(arr);
-            int byteSize = arr.length * Unsafe.ARRAY_SHORT_INDEX_SCALE;
+            long byteSize = (long)arr.length * Unsafe.ARRAY_SHORT_INDEX_SCALE;
             MemoryScope scope = MemoryScope.createConfined(null, MemoryScope.DUMMY_CLEANUP_ACTION, null);
             return new OfShort(Unsafe.ARRAY_SHORT_BASE_OFFSET, arr, byteSize, defaultAccessModes(byteSize), scope);
         }
@@ -170,7 +170,7 @@ public abstract class HeapMemorySegmentImpl<H> extends AbstractMemorySegmentImpl
 
         public static MemorySegment fromArray(int[] arr) {
             Objects.requireNonNull(arr);
-            int byteSize = arr.length * Unsafe.ARRAY_INT_INDEX_SCALE;
+            long byteSize = (long)arr.length * Unsafe.ARRAY_INT_INDEX_SCALE;
             MemoryScope scope = MemoryScope.createConfined(null, MemoryScope.DUMMY_CLEANUP_ACTION, null);
             return new OfInt(Unsafe.ARRAY_INT_BASE_OFFSET, arr, byteSize, defaultAccessModes(byteSize), scope);
         }
@@ -194,7 +194,7 @@ public abstract class HeapMemorySegmentImpl<H> extends AbstractMemorySegmentImpl
 
         public static MemorySegment fromArray(long[] arr) {
             Objects.requireNonNull(arr);
-            int byteSize = arr.length * Unsafe.ARRAY_LONG_INDEX_SCALE;
+            long byteSize = (long)arr.length * Unsafe.ARRAY_LONG_INDEX_SCALE;
             MemoryScope scope = MemoryScope.createConfined(null, MemoryScope.DUMMY_CLEANUP_ACTION, null);
             return new OfLong(Unsafe.ARRAY_LONG_BASE_OFFSET, arr, byteSize, defaultAccessModes(byteSize), scope);
         }
@@ -218,7 +218,7 @@ public abstract class HeapMemorySegmentImpl<H> extends AbstractMemorySegmentImpl
 
         public static MemorySegment fromArray(float[] arr) {
             Objects.requireNonNull(arr);
-            int byteSize = arr.length * Unsafe.ARRAY_FLOAT_INDEX_SCALE;
+            long byteSize = (long)arr.length * Unsafe.ARRAY_FLOAT_INDEX_SCALE;
             MemoryScope scope = MemoryScope.createConfined(null, MemoryScope.DUMMY_CLEANUP_ACTION, null);
             return new OfFloat(Unsafe.ARRAY_FLOAT_BASE_OFFSET, arr, byteSize, defaultAccessModes(byteSize), scope);
         }
@@ -242,7 +242,7 @@ public abstract class HeapMemorySegmentImpl<H> extends AbstractMemorySegmentImpl
 
         public static MemorySegment fromArray(double[] arr) {
             Objects.requireNonNull(arr);
-            int byteSize = arr.length * Unsafe.ARRAY_DOUBLE_INDEX_SCALE;
+            long byteSize = (long)arr.length * Unsafe.ARRAY_DOUBLE_INDEX_SCALE;
             MemoryScope scope = MemoryScope.createConfined(null, MemoryScope.DUMMY_CLEANUP_ACTION, null);
             return new OfDouble(Unsafe.ARRAY_DOUBLE_BASE_OFFSET, arr, byteSize, defaultAccessModes(byteSize), scope);
         }
